feat: handle Decimal in Appender.bind_parameter
