+++ crates/duckdb/src/types/mod.rs
@@ -3,6 +3,7 @@
 //! a value was NULL (which gets translated to `None`).
 
 pub use self::{
+    decimal::to_duckdb_decimal,
     from_sql::{FromSql, FromSqlError, FromSqlResult},
     ordered_map::OrderedMap,
     string::DuckString,
@@ -25,6 +26,7 @@ mod url;
 mod value;
 mod value_ref;
 
+mod decimal;
 mod ordered_map;
 mod string;
